move: `logger.js` 를 `utils/log` 폴더로 이동
- 로거 스위치를 만들기 위한 사전작업

Ref: #18
