+++ src/classes/models/playerGameData.class.js
@@ -8,7 +8,7 @@ import {
 import { getGameAssetById, initializeSpells } from '../../utils/assets/getAssets.js';
 import CustomErr from '../../utils/error/customErr.js';
 import { ERR_CODES } from '../../utils/error/errCodes.js';
-import logger from '../../utils/logger.js';
+import logger from '../../utils/log/logger.js';
 import Unit from './unit.class.js';
 import { SPELL_TYPE } from '../../constants/assets.js'; // eslint-disable-line
