+++ src/events/onError.js
@@ -1,7 +1,7 @@
 import gameSessionManager from '../classes/managers/gameSessionManager.js';
 import userSessionManager from '../classes/managers/userSessionManager.js';
 import { handleErr } from '../utils/error/handlerErr.js';
-import logger from '../utils/logger.js';
+import logger from '../utils/log/logger.js';
 
 export const onError = (socket) => (err) => {
   try {
