+++ src/handlers/sync/locationNotification.js
@@ -8,7 +8,7 @@ import adjustPos from '../../utils/location/adjustPos.js';
 import isValidPos from '../../utils/location/isValidPos.js';
 import gameSessionManager from '../../classes/managers/gameSessionManager.js';
 import formatCoords from '../../utils/formatter/formatCoords.js';
-import logger from '../../utils/logger.js';
+import logger from '../../utils/log/logger.js';
 
 /**
  * **위치 동기화 핸들러**
